# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chores Your Own Adventure API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parties:
    get:
      operationId: listMyParties
      tags: [parties]
      summary: List parties the current user belongs to
      responses:
        "200":
          description: List of parties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartyOverview"
    post:
      operationId: createParty
      tags: [parties]
      summary: Create a new party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartyInput"
      responses:
        "201":
          description: Created party
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
  /parties/{partyId}:
    get:
      operationId: getParty
      tags: [parties]
      summary: Get party details
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Party details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
    patch:
      operationId: updateParty
      tags: [parties]
      summary: Update party settings
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartyUpdate"
      responses:
        "200":
          description: Updated party
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
    delete:
      operationId: deleteParty
      tags: [parties]
      summary: Delete a party (founder only)
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Party deleted
  /parties/{partyId}/members:
    get:
      operationId: listPartyMembers
      tags: [parties]
      summary: List party members
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartyMember"
    post:
      operationId: addKidMember
      tags: [parties]
      summary: Add a kid adventurer to a party (leader only)
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KidMemberInput"
      responses:
        "201":
          description: Kid member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyMember"
  /parties/{partyId}/members/{userId}:
    patch:
      operationId: updatePartyMember
      tags: [parties]
      summary: Update member role or PIN reset
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
        - name: userId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyMember"
    delete:
      operationId: removePartyMember
      tags: [parties]
      summary: Remove member from party
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Member removed
  /parties/{partyId}/invite:
    post:
      operationId: createInvite
      tags: [parties]
      summary: Create an invite link or code
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /parties/{partyId}/pause-routines:
    post:
      operationId: pauseRoutineQuests
      tags: [parties]
      summary: Pause or resume all routine quests for a party
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PauseRoutinesInput"
      responses:
        "200":
          description: Paused/resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /parties/{partyId}/summon-cat-foley:
    post:
      operationId: summonCatFoley
      tags: [catFoley]
      summary: Summon Cat Foley (leader only)
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Cat Foley summoned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatFoleyVisit"
        "409":
          description: Already visiting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /parties/{partyId}/transfer-founder:
    post:
      operationId: transferFounder
      tags: [parties]
      summary: Transfer founder status to another leader
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferFounderInput"
      responses:
        "200":
          description: Founder transferred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /parties/{partyId}/leave:
    post:
      operationId: leaveParty
      tags: [parties]
      summary: Leave a party
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Left party
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /parties/join:
    post:
      operationId: joinParty
      tags: [parties]
      summary: Join a party via invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinPartyInput"
      responses:
        "200":
          description: Joined party
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"

  # ─── CHARACTER ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ──────────────────────────────────────────────────────────
...
    PartyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    PartyUpdate:
      type: object
      properties:
        name:
          type: string
        routinesPaused:
          type: boolean
        routinesPausedUntil:
          type: ["string", "null"]
    Party:
      type: object
      required: [id, name, householdCode, founderId, memberCount, routinesPaused, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        householdCode:
          type: string
        founderId:
          type: number
        memberCount:
          type: number
        routinesPaused:
          type: boolean
        routinesPausedUntil:
          type: ["string", "null"]
        partyGoldReserve:
          type: number
        myRole:
          type: ["string", "null"]
          enum: [founder, leader, adventurer, null]
        createdAt:
          type: string
    PartyOverview:
      type: object
      required: [id, name, householdCode, myRole, memberCount]
      properties:
        id:
          type: number
        name:
          type: string
        householdCode:
          type: string
        myRole:
          type: string
          enum: [founder, leader, adventurer]
        memberCount:
          type: number
        partyGoldReserve:
          type: number
        routinesPaused:
          type: boolean
    PartyMember:
      type: object
      required: [userId, displayName, role, userType, level]
      properties:
        userId:
          type: number
        displayName:
          type: string
        adventurerName:
          type: ["string", "null"]
        role:
          type: string
          enum: [founder, leader, adventurer]
        userType:
          type: string
          enum: [adult, kid]
        level:
          type: number
        species:
          type: ["string", "null"]
        class:
          type: ["string", "null"]
        joinedAt:
          type: string
    KidMemberInput:
      type: object
      required: [displayName, pin]
      properties:
        displayName:
          type: string
          minLength: 1
        pin:
          type: string
          minLength: 4
          maxLength: 4
    MemberUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [leader, adventurer]
        resetPin:
          type: string
          minLength: 4
          maxLength: 4
        adventurerName:
          type: string
    InviteInput:
      type: object
      required: [inviteeRole]
      properties:
        inviteeRole:
          type: string
          enum: [leader, adventurer]
    Invite:
      type: object
      required: [id, code, inviteeRole, expiresAt]
      properties:
        id:
          type: number
        code:
          type: string
        inviteeRole:
          type: string
        expiresAt:
          type: string
    JoinPartyInput:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
    PauseRoutinesInput:
      type: object
      required: [pause]
      properties:
        pause:
          type: boolean
        resumeOn:
          type: ["string", "null"]
    TransferFounderInput:
      type: object
      required: [newFounderUserId]
      properties:
        newFounderUserId:
          type: number

    # ─── CHARACTER SCHEMAS ─────────────────────────────────────────────────────
...
    ShopItem:
      type: object
      required: [id, name, category, goldPrice, minLevel, active]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [weapon, off_hand, outfit, head, pet, pet_accessory, background, effect]
        goldPrice:
          type: number
        minLevel:
          type: number
        minLegendaryCompletions:
          type: number
        isLocked:
          type: boolean
        isOwned:
          type: boolean
        isCatFoleyExclusive:
          type: boolean
        neverReturn:
          type: boolean
        active:
          type: boolean
        isEvolvingPet:
          type: boolean
        evolutionStages:
          type: array
          items:
            type: string
        assetKey:
          type: ["string", "null"]
        emoji:
          type: ["string", "null"]
...
    CatFoleyVisit:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
        visitId:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        dialogue:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/ShopItem"

    # ─── SCHOOL CALENDAR SCHEMAS ───────────────────────────────────────────────
...