# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Chores Your Own Adventure API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberUpdateRole } from './memberUpdateRole';

export interface MemberUpdate {
  role?: MemberUpdateRole;
  /**
     * @minLength 4
     * @maxLength 4
     */
  resetPin?: string;
  adventurerName?: string;
}