# lib/api-zod/src/generated/types/party.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Chores Your Own Adventure API
 * OpenAPI spec version: 0.1.0
 */
import type { PartyMyRole } from './partyMyRole';

export interface Party {
  id: number;
  name: string;
  householdCode: string;
  founderId: number;
  memberCount: number;
  routinesPaused: boolean;
  /** @nullable */
  routinesPausedUntil?: string | null;
  partyGoldReserve?: number;
  /** @nullable */
  myRole?: PartyMyRole;
  createdAt: string;
}